Voting is completely finished
User can now configure the vote to be a two thirds super majority OR a simple majority
vote can also now properly be reset
countries can vote in favor, against, or abstain if they said they were JUST present. countries that are PV cannot abstain. if the vote reaches >= majority, it shows a screen telling the user their vote passed. if they do not reach that amount, it shows the user a vote failed screen